// src/app/openapi/model/leave.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Leave { 
    id?: string;
    startDate?: string;
    endDate?: string;
    reason?: string;
    status?: Leave.StatusEnum;
}
export namespace Leave {
    export type StatusEnum = 'PENDING' | 'APPROVED' | 'REJECTED';
    export const StatusEnum = {
        Pending: 'PENDING' as StatusEnum,
        Approved: 'APPROVED' as StatusEnum,
        Rejected: 'REJECTED' as StatusEnum
    };
}